repositories_used/system-4b-augmentation-phase-1/web service/requirements.txt
dlib==19.24.0
scipy==1.6.2
Pillow==9.2.0
opencv_contrib_python==4.5.5.64
numpy==1.21.5
